+++ Source/Extensions/Blazorise.Charts/Options/Axis.cs
@@ -1,0 +1,63 @@
+﻿#region Using directives
+using System.Runtime.Serialization;
+using System.Text.Json.Serialization;
+#endregion
+
+namespace Blazorise.Charts
+{
+    [DataContract]
+    public class Axis
+    {
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public string Id { get; set; }
+
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public string Type { get; set; }
+
+        [DataMember]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public bool? Display { get; set; } = true;
+
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public AxisTicks Ticks { get; set; }
+
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public AxisGridLines GridLines { get; set; }
+
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public AxisScaleLabel ScaleLabel { get; set; }
+
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public bool? Stacked { get; set; }
+
+        /// <summary>
+        /// Manually set width of each bar in pixels. If not set, the base sample widths are calculated automatically so that they take the full available widths without overlap. Then, the bars are sized using barPercentage and categoryPercentage.
+        /// </summary>
+        /// <remarks>
+        /// NOTE: used only on Bar chart!
+        /// </remarks>
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public string BarThickness { get; set; }
+
+        /// <summary>
+        /// Set this to ensure that bars are not sized thicker than this.
+        /// </summary>
+        /// <remarks>
+        /// NOTE: used only on Bar chart!
+        /// </remarks>
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public int? MaxBarThickness { get; set; }
+
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public string Position { get; set; }
+    }
+}
