+++ Source/Extensions/Blazorise.Charts/Options/Legend.cs
@@ -1,0 +1,39 @@
+﻿#region Using directives
+using System.Runtime.Serialization;
+using System.Text.Json.Serialization;
+#endregion
+
+namespace Blazorise.Charts
+{
+    [DataContract]
+    public class Legend
+    {
+        /// <summary>
+        /// Is the legend shown.
+        /// </summary>
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public bool? Display { get; set; } = true;
+
+        /// <summary>
+        /// Marks that this box should take the full width of the canvas (pushing down other boxes). This is unlikely to need to be changed in day-to-day use.
+        /// </summary>
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public bool? FullWidth { get; set; } = true;
+
+        /// <summary>
+        /// Legend will show datasets in reverse order.
+        /// </summary>
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public bool? Reverse { get; set; } = false;
+
+        /// <summary>
+        /// Options to change legend labels.
+        /// </summary>
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public LegendLabels Labels { get; set; }
+    }
+}
