+++ Source/Extensions/Blazorise.Charts/Options/Specifics/DoughnutChartOptions.cs
@@ -1,0 +1,18 @@
+﻿#region Using directives
+using System.Runtime.Serialization;
+using System.Text.Json.Serialization;
+#endregion
+
+namespace Blazorise.Charts
+{
+    [DataContract]
+    public class DoughnutChartOptions : PieChartOptions
+    {
+        /// <summary>
+        /// The percentage of the chart that is cut out of the middle.
+        /// </summary>
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public new int? CutoutPercentage { get; set; } = 50;
+    }
+}
