+++ Source/Extensions/Blazorise.Charts/Options/Specifics/PolarAreaChartOptions.cs
@@ -1,0 +1,19 @@
+﻿#region Using directives
+using System;
+using System.Runtime.Serialization;
+using System.Text.Json.Serialization;
+#endregion
+
+namespace Blazorise.Charts
+{
+    [DataContract]
+    public class PolarAreaChartOptions : ChartOptions
+    {
+        /// <summary>
+        /// Starting angle to draw arcs for the first item in a dataset.
+        /// </summary>
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public double? StartAngle { get; set; } = -0.5 * Math.PI;
+    }
+}
