+++ Source/Extensions/Blazorise.Charts/Options/Tooltips.cs
@@ -1,0 +1,18 @@
+﻿#region Using directives
+using System.Runtime.Serialization;
+using System.Text.Json.Serialization;
+#endregion
+
+namespace Blazorise.Charts
+{
+    [DataContract]
+    public class Tooltips
+    {
+        /// <summary>
+        /// Are on-canvas tooltips enabled.
+        /// </summary>
+        [DataMember( EmitDefaultValue = false )]
+        [JsonIgnore( Condition = JsonIgnoreCondition.WhenWritingNull )]
+        public bool? Enabled { get; set; } = true;
+    }
+}
